openssl_lib_mock: Forcefully destroy mock
It seems that the static (de)initialization nightmare
hits us with the OpenSSLMockManager class that implements
the singleton. Gmock might not see the (last) mock object
destroyed, does not check the expected calls and returns an
error.

In order to avoid this, explicitly destroy the mock interface
at the end of a test case.